Fix Alexandr's comments
