+++ modules/metastorage/pom.xml
@@ -40,7 +40,7 @@
 
         <dependency>
             <groupId>org.apache.ignite</groupId>
-            <artifactId>ignite-join</artifactId>
+            <artifactId>ignite-cluster-management</artifactId>
         </dependency>
 
         <dependency>
